+++ Sources/BLENearbyExchangeCore/BLE/Central/BLECentralManager.swift
@@ -6,7 +6,7 @@ import Foundation
 final class BLECentralManager: NSObject, BLECentralInterface {
   private let configuration: NearbyExchange.Configuration
 
-  private var nonce: UInt64?
+  private var nonce: UInt32?
   private var manager: CBMCentralManager!
   private var ranger: ProximityRanger
 
@@ -47,7 +47,7 @@ final class BLECentralManager: NSObject, BLECentralInterface {
     )
   }
 
-  func startScanning(nonce: UInt64) {
+  func startScanning(nonce: UInt32) {
     self.nonce = nonce
 
     manager.scanForPeripherals(
